return `EVT_VARIANT` and `EVT_RPC_LOGIN` to Winevt.java change WString to String
